requirements.txt
alembic==0.9.6
aniso8601==1.3.0
arrow==0.12.0
attrs==17.3.0
blessings==1.6
boto3==1.4.4
botocore==1.8.5
bpython==0.16
celery==4.1.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.2
curtsies==0.2.11
docutils==0.14
Flask==0.12.1
Flask-HTTPAuth==3.2.3
Flask-Limiter==0.9.4
Flask-Migrate==2.0.3
flask-restplus==0.11.0
Flask-Script==2.0.5
flask-shell-bpython==0.0.1
Flask-SQLAlchemy==2.2
greenlet==0.4.12
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
limits==1.2.1
Mako==1.0.7
MarkupSafe==1.0
pluggy==0.6.0
py==1.5.2
Pygments==2.2.0
PySocks==1.6.8
pytest==3.3.0
pytest-cov==2.5.1
pytest-flask==0.10.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
SQLAlchemy==1.1.15
swagger-spec-validator==2.1.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
slackclient==1.1.0
Flask-Admin==1.5.1
Flask-Login==0.4.1
timezones==2.0.1
flask_wtf==0.14.2
flask_migrate
